requirements.txt
# Core dependencies
boto3>=1.34.0
sagemaker>=2.200.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
PyYAML>=6.0

# AWS utilities
awscli>=1.32.0

# Data processing
pyarrow>=14.0.0
fastparquet>=2023.10.0

# Monitoring
evidently>=0.4.0  # Data drift detection

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
